requirements.txt
Django==1.7.4

#Admin
django-filebrowser==3.5.3
django-grappelli==2.6.3
feedparser==5.1.3
pillow==2.3.0
django-localflavor==1.1

#Static hosting
django-pipeline==1.4.6
libsass==0.8.2
cssmin==0.2.0
slimit==0.8.1
jsmin==2.0.11
manifesto==0.5

#DB
psycopg2==2.5.2

#User Accounts
django-registration-redux==1.2
python-social-auth==0.1.23


#Utils
django-annoying==0.7.9
django-extensions==1.4.0
django-model-utils==2.0
django-braces==1.4.0
git+https://github.com/mikeorr/Unipath.git#egg=unipath
pytz==2013.9
easy-thumbnails==2.0.1
django-jsonfield==0.9.12


#REDIS CACHING
hiredis==0.1.2
redis==2.9.1
django-redis-cache==0.10.2
django-redis-sessions==0.4.0

## AWS
boto==2.24.0
django-storages-redux==1.2.2

#HEROKU
dj-database-url==0.2.2
#dj-static==0.0.5
-e git+https://github.com/gone/dj-static.git@5da00529c033afa8e65aef13555b9eaec6dfc388#egg=dj-static
django-toolbelt==0.0.1
ipython==1.1.0
waitress==0.8.9
djrill==1.3.0

#cool stuff
django-reversion==1.8.0

#OPTIONAL TOOLS
#whoosh
#django-haystack


#Money
stripe==1.12.0
